Added rob_rot2eul, changed conversion functions names (see description)
For greater clarity, in the conversion functions between homogeneous transformation matrices, rotation matrices, angles and quaternions, changed "r" to "rot" and "q" to "quat". For example: rob_q2r is now rob_quat2rot, rob_tr2r is now rob_tr2rot, and so on. "tr" is still the same (homogeneous transformation matrix), "rpy" is still the same for roll-pitch-yaw and "eul" is still the same for euler angles.